backend/movies_backend/pyproject.toml
[tool.poetry]
name = "movies-backend"
version = "0.1.0"
description = ""
authors = ["Vadim Titov <vadim.titov.hh@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pyyaml = "^6.0.2"
sqlalchemy = "^2.0.35"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
commitizen = "^3.29.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.11.2"
pylint = "^3.2.7"
black = "^24.8.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-local-badge = "^1.0.3"
httpx = "^0.27.2"
create-license = "^2.1.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[tool.black]
preview = true
line-length = 79
enable-unstable-feature = ["string_processing", "multiline_string_handling"]
